+++ Plugins/MsCrmTools.SynchronousEventOrderEditor/MsCrmTools.SynchronousEventOrderEditor.csproj
@@ -126,6 +126,7 @@
     <Compile Include="AppCode\PluginStep.cs" />
     <Compile Include="AppCode\SynchronousWorkflow.cs" />
     <Compile Include="AppCode\TreeViewHelper.cs" />
+    <Compile Include="AppCode\WorkflowHelper.cs" />
     <Compile Include="MainControl.cs">
       <SubType>UserControl</SubType>
     </Compile>
@@ -174,7 +175,7 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties BuildVersion_ConfigurationName="Release" BuildVersion_UpdateFileVersion="True" BuildVersion_UpdateAssemblyVersion="True" BuildVersion_BuildVersioningStyle="None.YearStamp.MonthStamp.DayStamp" />
+      <UserProperties BuildVersion_BuildVersioningStyle="None.YearStamp.MonthStamp.DayStamp" BuildVersion_UpdateAssemblyVersion="True" BuildVersion_UpdateFileVersion="True" BuildVersion_ConfigurationName="Release" />
     </VisualStudio>
   </ProjectExtensions>
   <PropertyGroup>
