+++ packages/repositories.config
@@ -1,7 +1,28 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <repositories>
+  <repository path="..\Plugins\DamSim.ViewTransferTool\packages.config" />
+  <repository path="..\Plugins\GapConsulting.PowerBIOptionSetAssistant\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.AttributeBulkUpdater\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.AuditCenter\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.ChartManager\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.FetchXmlTester\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.FlsBulkUpdater\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.FormRelated\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.Iconator\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.MetadataBrowser\packages.config" />
   <repository path="..\Plugins\MsCrmTools.MetadataDocumentGenerator\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.SampleTool\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.ScriptsFinder\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.SecurityRelated\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.SiteMapEditor\packages.config" />
   <repository path="..\Plugins\MsCrmTools.SolutionImport\packages.config" />
+  <repository path="..\Plugins\MscrmTools.SyncFilterManager\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.SynchronousEventOrderEditor\packages.config" />
   <repository path="..\Plugins\MsCrmTools.Translator\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.UserSettingsUtility\packages.config" />
+  <repository path="..\Plugins\MsCrmTools.ViewLayoutReplicator\packages.config" />
   <repository path="..\Plugins\MsCrmTools.WebResourcesManager\packages.config" />
+  <repository path="..\XrmToolBox.AutoUpdater\packages.config" />
+  <repository path="..\XrmToolBox.Extensibility\packages.config" />
+  <repository path="..\XrmToolBox\packages.config" />
 </repositories>
